@@ -1,31 +1,59 @@
 {
-    "name": "italia/spid-php-lib",
-    "description": "PHP package for SPID authentication",
-    "type": "library",
-    "license": "BSD-3-Clause",
-    "authors": [{
-            "name": "Lorenzo Cattaneo",
-            "email": "development@lorenzocattaneo.com"
-        },
-        {
-            "name": "Paolo Greppi",
-            "email": "paolo.greppi@libpf.com"
-        }
-    ],
-    "require": {
-        "robrichards/xmlseclibs": "^3.0",
-        "php": "^7.4 || ^8.0"
+  "name": "italia/spid-php-lib",
+  "description": "PHP package for SPID authentication",
+  "type": "library",
+  "license": "BSD-3-Clause",
+  "authors": [
+    {
+      "name": "Lorenzo Cattaneo",
+      "email": "development@lorenzocattaneo.com"
     },
-    "require-dev": {
-        "squizlabs/php_codesniffer": "^3.3",
-        "phpunit/phpunit": "^9.5"
-    },
-    "autoload": {
-        "psr-4": {
-            "Italia\\Spid\\": "src/"
-        }
-    },
-    "archive": {
-        "exclude": ["/example", "/tests"]
+    {
+      "name": "Paolo Greppi",
+      "email": "paolo.greppi@libpf.com"
     }
+  ],
+  "require": {
+    "robrichards/xmlseclibs": "^3.0",
+    "php": "^8.0",
+    "ext-openssl": "*",
+    "ext-session": "*",
+    "ext-dom": "*",
+    "ext-random": "*",
+    "ext-zlib": "*",
+    "ext-filter": "*",
+    "ext-simplexml": "*"
+  },
+  "require-dev": {
+    "vimeo/psalm": "^5.15",
+    "phpstan/phpstan": "^1.10",
+    "symplify/easy-coding-standard": "^12.0",
+    "maglnet/composer-require-checker": "^4.7"
+  },
+  "autoload": {
+    "psr-4": {
+      "Italia\\Spid\\": "src/"
+    }
+  },
+  "archive": {
+    "exclude": [
+      "/example",
+      "/tests"
+    ]
+  },
+  "scripts": {
+    "ecs": "ecs check",
+    "ecs-fix": "ecs check --fix",
+    "phpstan": "phpstan analyse",
+    "psalm": "psalm",
+    "require-checker": "composer-require-checker check --config-file=require-checker-config.json",
+    "phpstan-baseline": "phpstan --generate-baseline=phpstan-baseline.neon",
+    "psalm-baseline": "psalm --set-baseline=psalm-baseline.xml",
+    "test-static": [
+      "@require-checker",
+      "@ecs",
+      "@phpstan",
+      "@psalm"
+    ]
+  }
 }
