@@ -19,7 +19,7 @@ interface IdpInterface
     // $returnTo: return url
     // $shouldRedirect: tells if the function should emit headers and redirect to login URL or return the URL as string
     // returns and empty string if $shouldRedirect = true, the login URL otherwhise
-    public function authnRequest($ass, $attr, $level, $returnTo = null, $shouldRedirect = true) : string;
+    public function authnRequest($ass, $attr, $level, $returnTo = null, $shouldRedirect = true): string;
 
     // generate a LogoutRequest
     // $session: the currently active login session
@@ -28,7 +28,7 @@ interface IdpInterface
     // $returnTo: return url
     // $shouldRedirect: tells if the function should emit headers and redirect to login URL or return the URL as string
     // returns and empty string if $shouldRedirect = true, the logout URL otherwhise
-    public function logoutRequest(Session $session, $slo, $binding, $returnTo = null, $shouldRedirect = true) : string;
+    public function logoutRequest(Session $session, $slo, $binding, $returnTo = null, $shouldRedirect = true): string;
 
     //generates a logoutResponse in response to an Idp initiated logout request
     public function logoutResponse(): string;
