@@ -31,7 +31,7 @@ XML;
         $this->xml = $xml;
     }
 
-    public function redirectUrl($redirectTo = null) : string
+    public function redirectUrl($redirectTo = null): string
     {
         $location = parent::getBindingLocation(Settings::BINDING_REDIRECT, 'SLO');
         if (is_null($this->xml)) {
@@ -40,7 +40,7 @@ XML;
         return parent::redirect($location, $redirectTo);
     }
 
-    public function httpPost($redirectTo = null) : string
+    public function httpPost($redirectTo = null): string
     {
         $location = parent::getBindingLocation(Settings::BINDING_POST, 'SLO');
         if (is_null($this->xml)) {
