@@ -3,9 +3,9 @@
 namespace Italia\Spid\Spid\Saml\Out;
 
 use Italia\Spid\Spid\Interfaces\RequestInterface;
-use Italia\Spid\Spid\Saml\Settings;
 use Italia\Spid\Spid\Saml\Idp;
 use Italia\Spid\Spid\Saml\In\LogoutRequest;
+use Italia\Spid\Spid\Saml\Settings;
 use Italia\Spid\Spid\Saml\SignatureUtils;
 
 class LogoutResponse extends Base implements RequestInterface
@@ -31,7 +31,7 @@ XML;
         $this->xml = $xml;
     }
 
-    public function redirectUrl($redirectTo = null) : string
+    public function redirectUrl($redirectTo = null): string
     {
         $location = parent::getBindingLocation(Settings::BINDING_REDIRECT, 'SLO');
         if (is_null($this->xml)) {
@@ -40,7 +40,7 @@ XML;
         return parent::redirect($location, $redirectTo);
     }
 
-    public function httpPost($redirectTo = null) : string
+    public function httpPost($redirectTo = null): string
     {
         $location = parent::getBindingLocation(Settings::BINDING_POST, 'SLO');
         if (is_null($this->xml)) {
