@@ -17,7 +17,7 @@ class Session
         }
         foreach ($values as $key => $value) {
             if (property_exists($this, $key)) {
-                $this->{$key} =  $value;
+                $this->{$key} = $value;
             }
         }
     }
